data/5.Targets_NetworkDistance/OMA.2.5.0/hog_bottom_up/requirements.txt
numpy
biopython
lxml
